requirements.txt
fastapi~=0.115.14
uvicorn

pydantic~=2.11.7
SQLAlchemy~=2.0.41
bcrypt~=4.1.2
PyJWT~=2.8.0
email-validator~=2.1.0
psycopg2-binary~=2.9.9